FLSAM.control

Creates an FLSAM.control that can be used to control an FLSAM assessment

FLSAM.control(stck, tun, default="full",sam.binary="missing")

Arguments

stck

An FLStock object

tun

An FLIndices object

default

Specifies how the control object should be populated. Default option is "full", which takes a reasonable first guess at a configuration. Any other value will populate all matrices with NA values.

sam.binary

Path and name of the SAM executable file to run

Details

An FLSAM assessment requires three elements - an FLStock object, an FLIndices object and an FLSAM.control object. The FLSAM.control object contains the configuration options for the model, and is therefore key to the process. However, there are a great many configuration options and therefore setting up the control object is a relatively complex process. The purpose of the FLSAM.control function is to create a template (possibly containing sensible first guesses at a configuration) object that can then be populated by the user according to their wishes.

The key concept with the configuration of a SAM model is that some parameters are bound together, so that the model fits just a single parameter, but uses it in multiple places e.g. by using a single parameter for the catchabilites of ages 5-9 in a survey. We refer to this concept as "binding" the parameters together.

Bindings are expressed in a matrix format in the FLSAM.control object. The rows of the matrix correspond to the fleets, whilst the column correspond to the ages - it is important to remember that while all age-fleet combinations are represented here, not all are valid i.e. a fleet may not contain information on all ages. Bindings are specified with integers - the integers indicate which parameter should be used in the specific fleet-age context represented by that point in the matrix. Repeating a number means that all those fleet-age combinations with the same number will be fitted with just one parameter i.e. bound together. Parameters that are not to be fitted, or are turned off, are specified by setting the value to NA.

FLSAM.control can populate the template in two different ways, as indicated by the "default" argument. If default="full", a reasonable first guess at a configuration is supplied - this consists of linear catchability models for the surveys and a fixed selectivity pattern in the fishery. Alternatively, if "default" is set to any other value, all values are set to NA and therefore turned "off".

Earlier versions of FLSAM contained the executable (binary) file within the package distribution. However, this version is designed to interface with customised versions of the sam model, which can be modified and compiled locally. The sam.binary slot in the FLSAM.control object can be used to specify the location of this executable file. If the file is not specified, FLSAM reverts to the versions stored within the package - however, in the future, these binaries may be removed.

Value

Returns an FLSAM.control object with the following slots

Warning

It is important to realise that the default configuration returned by FLSAM.control is almost certaintly not the best configuration of your model - it is merely a first guess setup for your convenience. Similarly, there is no guarantee that this model will lead to a converged solution. It is the users responsibility to find the most appropriate "best" configuration of the model.

See also

FLSAM

Examples

#Load data
library(FLSAM)
data(NSH)
#Create an object
ctrl <- FLSAM.control(NSH,NSH.tun)
ctrl

#> An object of class "FLSAM.control"
#> Slot "name":
#> character(0)
#> 
#> Slot "desc":
#> character(0)
#> 
#> Slot "range":
#>       min       max plusgroup   minyear   maxyear   minfbar   maxfbar 
#>         0         8         8      1992      2012         2         6 
#> 
#> Slot "fleets":
#>   catch    SCAI   HERAS IBTS-Q1   IBTS0 
#>       0       3       2       2       2 
#> 
#> Slot "plus.group":
#> plusgroup 
#>      TRUE 
#> 
#> Slot "states":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    1  1  1  1  1  1  1  1  1
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "logN.vars":
#> 0 1 2 3 4 5 6 7 8 
#> 1 1 1 1 1 1 1 1 1 
#> 
#> Slot "catchabilities":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch   NA NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA  1  2  3  4  5  6  7  8
#>   IBTS-Q1 NA  9 NA NA NA NA NA NA NA
#>   IBTS0   10 NA NA NA NA NA NA NA NA
#> 
#> Slot "power.law.exps":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch   NA NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "f.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    1  1  1  1  1  1  1  1  1
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "obs.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    1  2  3  4  5  6  7  8  9
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA 10 11 12 13 14 15 16 17
#>   IBTS-Q1 NA 18 NA NA NA NA NA NA NA
#>   IBTS0   19 NA NA NA NA NA NA NA NA
#> 
#> Slot "srr":
#> [1] 0
#> 
#> Slot "cor.F":
#> [1] FALSE
#> 
#> Slot "nohess":
#> [1] FALSE
#> 
#> Slot "timeout":
#> [1] 3600
#> 
#> Slot "sam.binary":
#> character(0)
#> 

#Bind HERAS catchabilities on ages 5:8
ctrl@catchabilities["HERAS",as.character(5:8)] <- 11
#Alternatively can also use update functionality
ctrl@catchabilities["HERAS",as.character(5:8)] <- 101
ctrl <- update(ctrl)
#See example of full control object
data(NSH.sam)
NSH.ctrl

#> An object of class "FLSAM.control"
#> Slot "name":
#> [1] "Final Assessment"
#> 
#> Slot "desc":
#> character(0)
#> 
#> Slot "range":
#>       min       max plusgroup   minyear   maxyear   minfbar   maxfbar 
#>         0         8         8      1992      2012         2         6 
#> 
#> Slot "fleets":
#>   catch    SCAI   HERAS IBTS-Q1   IBTS0 
#>       0       3       2       2       2 
#> 
#> Slot "plus.group":
#> plusgroup 
#>      TRUE 
#> 
#> Slot "states":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    1  2  3  4  5  6  7  8  8
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "logN.vars":
#> 0 1 2 3 4 5 6 7 8 
#> 1 2 2 2 2 2 2 2 2 
#> 
#> Slot "catchabilities":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch   NA NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA  3  3  4  4  5  5  5  5
#>   IBTS-Q1 NA  1 NA NA NA NA NA NA NA
#>   IBTS0    2 NA NA NA NA NA NA NA NA
#> 
#> Slot "power.law.exps":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch   NA NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "f.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    1  1  2  2  3  3  4  4  4
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "obs.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    3  4  4  4  4  4  5  5  5
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA  6  7  7  7  7  8  8  8
#>   IBTS-Q1 NA  1 NA NA NA NA NA NA NA
#>   IBTS0    2 NA NA NA NA NA NA NA NA
#> 
#> Slot "srr":
#> [1] 0
#> 
#> Slot "cor.F":
#> [1] FALSE
#> 
#> Slot "nohess":
#> [1] FALSE
#> 
#> Slot "timeout":
#> [1] 3600
#> 
#> Slot "sam.binary":
#> character(0)
#>