Drop an element from an FLSAM.control object

Drops multiple surveys or age groups from an FLSAM.control object

drop.from.control(object, fleets="missing",ages="missing")

Arguments

object

An FLSAM.control object

fleets

A vector containing the names of the fleets to be dropped

ages

A vector containing the names of the age groups to be dropped

Details

This function can be used to modify an FLSAM.control object by removing one or more fleets, or one or more age groups from the object. Changes in the sequence of bindings that may arise due to the removal of a fleet or age group are corrected for using the update function.

Value

An FLSAM.control object

See also

update, FLSAM.control

Examples

#Load assessment
library(FLSAM)
data(NSH.sam)
#NSH.ctrl before dropping
NSH.ctrl

#> An object of class "FLSAM.control"
#> Slot "name":
#> [1] "Final Assessment"
#> 
#> Slot "desc":
#> character(0)
#> 
#> Slot "range":
#>       min       max plusgroup   minyear   maxyear   minfbar   maxfbar 
#>         0         8         8      1992      2012         2         6 
#> 
#> Slot "fleets":
#>   catch    SCAI   HERAS IBTS-Q1   IBTS0 
#>       0       3       2       2       2 
#> 
#> Slot "plus.group":
#> plusgroup 
#>      TRUE 
#> 
#> Slot "states":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    1  2  3  4  5  6  7  8  8
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "logN.vars":
#> 0 1 2 3 4 5 6 7 8 
#> 1 2 2 2 2 2 2 2 2 
#> 
#> Slot "catchabilities":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch   NA NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA  3  3  4  4  5  5  5  5
#>   IBTS-Q1 NA  1 NA NA NA NA NA NA NA
#>   IBTS0    2 NA NA NA NA NA NA NA NA
#> 
#> Slot "power.law.exps":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch   NA NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "f.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    1  1  2  2  3  3  4  4  4
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "obs.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    3  4  4  4  4  4  5  5  5
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA  6  7  7  7  7  8  8  8
#>   IBTS-Q1 NA  1 NA NA NA NA NA NA NA
#>   IBTS0    2 NA NA NA NA NA NA NA NA
#> 
#> Slot "srr":
#> [1] 0
#> 
#> Slot "cor.F":
#> [1] FALSE
#> 
#> Slot "nohess":
#> [1] FALSE
#> 
#> Slot "timeout":
#> [1] 3600
#> 
#> Slot "sam.binary":
#> character(0)
#> 

#Drop MLAI index
drop.from.control(NSH.ctrl,fleets="MLAI")

#> An object of class "FLSAM.control"
#> Slot "name":
#> [1] "Final Assessment"
#> 
#> Slot "desc":
#> character(0)
#> 
#> Slot "range":
#>       min       max plusgroup   minyear   maxyear   minfbar   maxfbar 
#>         0         8         8      1992      2012         2         6 
#> 
#> Slot "fleets":
#>   catch    SCAI   HERAS IBTS-Q1   IBTS0 
#>       0       3       2       2       2 
#> 
#> Slot "plus.group":
#> plusgroup 
#>      TRUE 
#> 
#> Slot "states":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    1  2  3  4  5  6  7  8  8
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "logN.vars":
#> 0 1 2 3 4 5 6 7 8 
#> 1 2 2 2 2 2 2 2 2 
#> 
#> Slot "catchabilities":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch   NA NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA  3  3  4  4  5  5  5  5
#>   IBTS-Q1 NA  1 NA NA NA NA NA NA NA
#>   IBTS0    2 NA NA NA NA NA NA NA NA
#> 
#> Slot "power.law.exps":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch   NA NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "f.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    1  1  2  2  3  3  4  4  4
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA NA
#> 
#> Slot "obs.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7  8
#>   catch    3  4  4  4  4  4  5  5  5
#>   SCAI    NA NA NA NA NA NA NA NA NA
#>   HERAS   NA  6  7  7  7  7  8  8  8
#>   IBTS-Q1 NA  1 NA NA NA NA NA NA NA
#>   IBTS0    2 NA NA NA NA NA NA NA NA
#> 
#> Slot "srr":
#> [1] 0
#> 
#> Slot "cor.F":
#> [1] FALSE
#> 
#> Slot "nohess":
#> [1] FALSE
#> 
#> Slot "timeout":
#> [1] 3600
#> 
#> Slot "sam.binary":
#> character(0)
#> 

#Drop ages 8 and 9
drop.from.control(NSH.ctrl,ages=8:9)

#> An object of class "FLSAM.control"
#> Slot "name":
#> [1] "Final Assessment"
#> 
#> Slot "desc":
#> character(0)
#> 
#> Slot "range":
#>       min       max plusgroup   minyear   maxyear   minfbar   maxfbar 
#>         0         8         8      1992      2012         2         6 
#> 
#> Slot "fleets":
#>   catch    SCAI   HERAS IBTS-Q1   IBTS0 
#>       0       3       2       2       2 
#> 
#> Slot "plus.group":
#> plusgroup 
#>      TRUE 
#> 
#> Slot "states":
#>          age
#> fleet      0  1  2  3  4  5  6  7
#>   catch    1  2  3  4  5  6  7  8
#>   SCAI    NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA
#> 
#> Slot "logN.vars":
#> 0 1 2 3 4 5 6 7 
#> 1 2 2 2 2 2 2 2 
#> 
#> Slot "catchabilities":
#>          age
#> fleet      0  1  2  3  4  5  6  7
#>   catch   NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA
#>   HERAS   NA  3  3  4  4  5  5  5
#>   IBTS-Q1 NA  1 NA NA NA NA NA NA
#>   IBTS0    2 NA NA NA NA NA NA NA
#> 
#> Slot "power.law.exps":
#>          age
#> fleet      0  1  2  3  4  5  6  7
#>   catch   NA NA NA NA NA NA NA NA
#>   SCAI    NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA
#> 
#> Slot "f.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7
#>   catch    1  1  2  2  3  3  4  4
#>   SCAI    NA NA NA NA NA NA NA NA
#>   HERAS   NA NA NA NA NA NA NA NA
#>   IBTS-Q1 NA NA NA NA NA NA NA NA
#>   IBTS0   NA NA NA NA NA NA NA NA
#> 
#> Slot "obs.vars":
#>          age
#> fleet      0  1  2  3  4  5  6  7
#>   catch    3  4  4  4  4  4  5  5
#>   SCAI    NA NA NA NA NA NA NA NA
#>   HERAS   NA  6  7  7  7  7  8  8
#>   IBTS-Q1 NA  1 NA NA NA NA NA NA
#>   IBTS0    2 NA NA NA NA NA NA NA
#> 
#> Slot "srr":
#> [1] 0
#> 
#> Slot "cor.F":
#> [1] FALSE
#> 
#> Slot "nohess":
#> [1] FALSE
#> 
#> Slot "timeout":
#> [1] 3600
#> 
#> Slot "sam.binary":
#> character(0)
#>