Compare model fit performance among the models specified based on log-likelihood ratio's. Different in log-likelihood is calculated and degrees of freedom are given. P values are specified assuming a chi-quared distribution.

lr.test(object,...,type="sequential")

Arguments

object

An FLSAM or FLSAMs object

…

In combination with an FLSAM object: additional objects of the same type (need at least 1)

type

The type of comparison to make. A "sequential" comparison compares object 1 against object 2, object 2 against object 3,... whilt a "first" comparison compares object 1 against object 2, object 1 against object 3,...

Value

A summary table of the test characteristics is returned

Warning!

Log-likelihood tests can only be used for nested models (i.e. model parameterisation is different) and are not appropriate for different input data or different model types.

See also

looi

Examples

#- Load the data
data(HERAS.sams)

#- Run lr ratio test on example FLSAMs object
lr.test(HERAS.sams)

#>                       statistics
#> models                 Comparison Neg. log likel # Parameters Likel difference
#>   1 Final Assessment   1 vs. 2    196.308        26           2.09            
#>   2 Final Assessment 2 2 vs. 3    198.396        22           0.68            
#>   3 Final Assessment 3 3 vs. 4    199.076        21           3.29            
#>   4 Final Assessment 4            202.364        20                           
#>                       statistics
#> models                 Degrees of freedom P value
#>   1 Final Assessment   4                  0.3827 
#>   2 Final Assessment 2 1                  0.2435 
#>   3 Final Assessment 3 1                  0.0103 
#>   4 Final Assessment 4                           

#- Run lr ratio test on individual FLSAM objects
lr.test(HERAS.sams[[1]],HERAS.sams[[2]],HERAS.sams[[3]])

#>                       statistics
#> models                 Comparison Neg. log likel # Parameters Likel difference
#>   1 Final Assessment   1 vs. 2    196.308        26           2.09            
#>   2 Final Assessment 2 2 vs. 3    198.396        22           0.68            
#>   3 Final Assessment 3            199.076        21                           
#>                       statistics
#> models                 Degrees of freedom P value
#>   1 Final Assessment   4                  0.3827 
#>   2 Final Assessment 2 1                  0.2435 
#>   3 Final Assessment 3