Parameter accessor functions for FLSAM

Returns parameters fitted by the FLSAM asessment model

params(object,...)
coef(object,...)
coefficients(object,...)

Arguments

object
An FLSAM or FLSAMs object.

...
Other arguments

Details

These accessor functions provide direct access to the values returned by the SAM model. However, interpretation of these parameters can be challenging and requires some knowledge of the underlying workings of the SAM model. For more access in a more user-friendly format, see the accessors help.

The following parameter accessor functions are currently supported:

params - returns all values estimated by the model, including derived parameters such as SSB, TSB, etc.

coef - returns the parameters fitted by the model, but not derived parameters such as SSB, TSB etc.

coefficients - an alias for coef

Accessor functions are provided for both FLSAM and FLSAMs objects.

Note that the naming of these parameters is somewhat obtuse. A brief description of these parameters follows:

logFpar: Log-transformed catchability of the numbers-at-age survey indices. The binding of these parameters are controlled in the 'catchabilities' slot of the FLSAM.control object.

logSdLogFsta: Log-transformed standard deviation of the random walk in the (log-transformed) fishering mortalities. Bindings are controlled in the 'f.vars' slot.

logSdLogN: Log-transformed standard deviation of the random walk in (log-transformed) N. Bindings are controlled in the logN.vars slot

logSdLogObs: Log-transformed observation variance for each data source. Bindings are controlled in the 'obs.vars' slot

logScaleSSB: Log-transformed catchability of the SSB indices

logSdSSB: Log-transformed observation variance of the SSB indices

rec_loga: Log-transformed 'a' parameter of the stock-recruitment model. Fitting of this parameter is controlled by the 'srr' slot

rec_logb: Log-transformed 'b' parameter of the stock-recruitment model. Fitting of this parameter is controlled by the 'srr' slot

rho: Correlation coefficient in the correlated random-walk in fishing mortality. Fitting of this parameter is controlled by the 'cor.F' slot

logPowSSB: Log-transformed power-law coefficient of the SSB catchability model.

U: The state variables. The order of the state variables is N(year 1),F(year 1, N(year 2), F(year 2), where N are the numbers-at-age state variables and F the fishing mortality state variables)

Not all parameters may be present in a fit, depending on the specific configuration of the FLSAM.control.

Value

The parameter accessor functions return a data.frame whoose length depends on the calling function. params() returns all values (both fitted and dervied parameters), whereas coef() and coefficients() only return the fitted parameters.

Columns returned are

object.name - In cases where an FLSAMs object is supplied, the name of the corresponding object is stored in this column. There is no requirement that FLSAMs object have unique names - however, extracting and processing data can be challenging if they do not!

index - An indexing number produced internally by SAM

name - The name of the parameter

value - The value of the parameter

std.dev - The standard deviation of the fitted parameter

See also

Examples

#Load data
library(FLSAM)
data(NSH.sam)
#Extract parameter
catchabilities(NSH.sam)

#>      fleet age        value         CV         lbnd         ubnd
#> 1     SCAI   0 1.311017e-02 0.13350653 1.009172e-02 1.703144e-02
#> 2    HERAS   1 8.627635e-01 0.09719681 7.131094e-01 1.043824e+00
#> 3    HERAS   2 9.851568e-01 0.08331534 8.367311e-01 1.159911e+00
#> 4    HERAS   3 1.138150e+00 0.08352438 9.662784e-01 1.340593e+00
#> 5    HERAS   4 1.189462e+00 0.08659227 1.003788e+00 1.409482e+00
#> 6    HERAS   5 1.299284e+00 0.09393059 1.080808e+00 1.561922e+00
#> 7    HERAS   6 1.408318e+00 0.10704643 1.141776e+00 1.737083e+00
#> 8    HERAS   7 1.524058e+00 0.13146136 1.177876e+00 1.971986e+00
#> 9    HERAS   8 1.524058e+00 0.13146136 1.177876e+00 1.971986e+00
#> 10 IBTS-Q1   1 1.522673e-04 0.09134139 1.273079e-04 1.821202e-04
#> 11   IBTS0   0 3.029939e-06 0.10529273 2.464943e-06 3.724438e-06

obs.var(NSH.sam)

#>           fleet age     value         CV      lbnd      ubnd
#> 6  catch unique   0 0.1912349 0.12320568 0.1502078 0.2434680
#> 9  catch unique   1 0.1912349 0.12320568 0.1502078 0.2434680
#> 7  catch unique   2 0.1912349 0.12320568 0.1502078 0.2434680
#> 8  catch unique   3 0.1912349 0.12320568 0.1502078 0.2434680
#> 4  catch unique   4 0.1912349 0.12320568 0.1502078 0.2434680
#> 5  catch unique   5 0.1912349 0.12320568 0.1502078 0.2434680
#> 2  catch unique   6 0.1912349 0.12320568 0.1502078 0.2434680
#> 3  catch unique   7 0.1912349 0.12320568 0.1502078 0.2434680
#> 1  catch unique   8 0.1912349 0.12320568 0.1502078 0.2434680
#> 10         SCAI   0 0.5249091 0.16381088 0.3807548 0.7236402
#> 12        HERAS   1 0.2605214 0.08241186 0.2216629 0.3061920
#> 11        HERAS   2 0.2605214 0.08241186 0.2216629 0.3061920
#> 17        HERAS   3 0.2605214 0.08241186 0.2216629 0.3061920
#> 14        HERAS   4 0.2605214 0.08241186 0.2216629 0.3061920
#> 18        HERAS   5 0.2605214 0.08241186 0.2216629 0.3061920
#> 15        HERAS   6 0.2605214 0.08241186 0.2216629 0.3061920
#> 16        HERAS   7 0.2605214 0.08241186 0.2216629 0.3061920
#> 13        HERAS   8 0.2605214 0.08241186 0.2216629 0.3061920
#> 19      IBTS-Q1   1 0.3067726 0.17790761 0.2164605 0.4347650
#> 20        IBTS0   0 0.3658268 0.18620768 0.2539641 0.5269613

ssb(NSH.sam)

#>     year     value         CV      lbnd      ubnd
#> 391 1992 1024480.4 0.07678608  881337.1 1190872.6
#> 392 1993  686221.7 0.08380875  582270.8  808730.6
#> 393 1994  775489.1 0.08479732  656742.0  915707.3
#> 394 1995  769030.5 0.09052400  644003.1  918330.7
#> 395 1996  918293.7 0.09282774  765535.0 1101534.7
#> 396 1997 1136845.0 0.09046812  952123.4 1357404.4
#> 397 1998 1343899.5 0.08476133 1138194.5 1586781.3
#> 398 1999 1445965.4 0.08648385 1220510.1 1713067.4
#> 399 2000 1387045.3 0.08464887 1174995.2 1637364.0
#> 400 2001 1886733.4 0.08380122 1600949.0 2223533.0
#> 401 2002 2251911.2 0.08316464 1913198.8 2650589.3
#> 402 2003 2274713.7 0.07924589 1947472.2 2656942.8
#> 403 2004 2246923.5 0.07962125 1922265.2 2626414.5
#> 404 2005 2012617.4 0.08468709 1704802.7 2376010.4
#> 405 2006 1577565.1 0.08528415 1334725.5 1864586.9
#> 406 2007 1280446.9 0.08913459 1075198.6 1524875.8
#> 407 2008 1339230.2 0.09033500 1121916.5 1598637.3
#> 408 2009 1686097.0 0.09168668 1408760.8 2018031.0
#> 409 2010 1743199.8 0.09693649 1441561.4 2107954.2
#> 410 2011 2200542.2 0.10122844 1804522.5 2683472.1
#> 411 2012 2247280.2 0.12436981 1761130.4 2867628.9

fbar(NSH.sam)

#>     year      value         CV       lbnd       ubnd
#> 412 1992 0.42266255 0.09360483 0.35181612 0.50777558
#> 413 1993 0.52292856 0.09150993 0.43706651 0.62565827
#> 414 1994 0.48825713 0.09450857 0.40569651 0.58761912
#> 415 1995 0.49398540 0.10201307 0.40446303 0.60332233
#> 416 1996 0.23447425 0.11141036 0.18847803 0.29169539
#> 417 1997 0.20442207 0.11252855 0.16396138 0.25486725
#> 418 1998 0.25461706 0.11107126 0.20480556 0.31654339
#> 419 1999 0.22261618 0.10748075 0.18032970 0.27481864
#> 420 2000 0.23384297 0.10488981 0.19038832 0.28721581
#> 421 2001 0.19861817 0.10485816 0.16171932 0.24393608
#> 422 2002 0.18455656 0.10321194 0.15075569 0.22593591
#> 423 2003 0.20689806 0.10024805 0.16999007 0.25181945
#> 424 2004 0.24524743 0.09844270 0.20221266 0.29744083
#> 425 2005 0.31657402 0.09762018 0.26144437 0.38332862
#> 426 2006 0.26255497 0.10099016 0.21540495 0.32002566
#> 427 2007 0.23267623 0.10742955 0.18849773 0.28720890
#> 428 2008 0.14783289 0.11153645 0.11880352 0.18395551
#> 429 2009 0.07745840 0.11564283 0.06174920 0.09716407
#> 430 2010 0.08182316 0.11665003 0.06510011 0.10284206
#> 431 2011 0.09768402 0.12580389 0.07633735 0.12499998
#> 432 2012 0.09777294 0.37335569 0.04703400 0.20324762

rec(NSH.sam)

#>     year    value        CV     lbnd      ubnd
#> 188 1992 58018537 0.1313998 44845306  75061381
#> 197 1993 48390196 0.1285922 37609492  62261174
#> 206 1994 34341049 0.1326573 26478499  44538311
#> 215 1995 49483998 0.1438629 37325548  65602948
#> 224 1996 42860629 0.1402732 32557840  56423692
#> 233 1997 27788062 0.1475876 20807944  37109692
#> 242 1998 17679573 0.1559988 13022157  24002727
#> 251 1999 66144238 0.1476299 49525276  88339946
#> 260 2000 46246427 0.1436452 34898350  61284616
#> 269 2001 81981185 0.1471774 61437605 109394152
#> 278 2002 42861860 0.1439789 32323149  56836635
#> 287 2003 19709009 0.1460348 14803261  26240503
#> 296 2004 25222541 0.1471925 18901490  33657483
#> 305 2005 22604866 0.1421042 17109622  29865065
#> 314 2006 26434005 0.1437846 19942107  35039259
#> 323 2007 24224957 0.1564599 17827140  32918827
#> 332 2008 27296142 0.1628126 19838660  37556939
#> 341 2009 37671892 0.1548334 27811252  51028677
#> 350 2010 32573646 0.1695605 23363266  45414986
#> 359 2011 32898140 0.1993223 22258957  48622566
#> 368 2012 27858359 0.3171078 14963319  51866044

#Power law parameters could be extracted as follows, but this will
#return an error in this case, as there are no power-law exponents in the model
#power.law.exps(NSH.sam) 

#And for FLSAMs
data(HERAS.sams)
ssb(HERAS.sams)

#>            name year     value         CV      lbnd      ubnd
#> 390    all.free 1992 1032006.5 0.08092920  880631.3 1209402.2
#> 391    all.free 1993  692846.6 0.08916750  581749.8  825159.7
#> 392    all.free 1994  773355.7 0.08960060  648798.3  921825.8
#> 393    all.free 1995  770355.6 0.09504940  639416.1  928108.8
#> 394    all.free 1996  927606.5 0.09747391  766288.5 1122884.9
#> 395    all.free 1997 1149316.5 0.09547671  953165.3 1385833.5
#> 396    all.free 1998 1328062.8 0.08889634 1115702.8 1580842.9
#> 397    all.free 1999 1425038.7 0.09117667 1191833.3 1703875.4
#> 398    all.free 2000 1383828.4 0.08933151 1161560.2 1648628.3
#> 399    all.free 2001 1882761.0 0.08826739 1583654.6 2238359.9
#> 400    all.free 2002 2244278.6 0.08778101 1889539.9 2665615.3
#> 401    all.free 2003 2273798.0 0.08377291 1929491.8 2679543.6
#> 402    all.free 2004 2239076.4 0.08407190 1898914.7 2640172.8
#> 403    all.free 2005 2006711.7 0.08881394 1686106.6 2388278.3
#> 404    all.free 2006 1588761.3 0.09034591 1330928.3 1896542.7
#> 405    all.free 2007 1290650.1 0.09508351 1071203.1 1555053.1
#> 406    all.free 2008 1353345.4 0.09629868 1120566.3 1634480.5
#> 407    all.free 2009 1712693.1 0.09714603 1415752.0 2071915.0
#> 408    all.free 2010 1781186.8 0.10171298 1459249.8 2174149.0
#> 409    all.free 2011 2298416.6 0.10908925 1855966.4 2846344.0
#> 410    all.free 2012 2346130.1 0.13957405 1784613.9 3084323.6
#> 388   four.lvls 1992 1040449.0 0.07943575  890438.1 1215732.0
#> 389   four.lvls 1993  700324.0 0.08705563  590467.1  830619.7
#> 3901  four.lvls 1994  790803.7 0.08736400  666350.8  938500.3
#> 3911  four.lvls 1995  783214.8 0.09315070  652513.4  940096.4
#> 3921  four.lvls 1996  934324.7 0.09527255  775175.8 1126147.9
#> 3931  four.lvls 1997 1156719.8 0.09266430  964607.7 1387093.1
#> 3941  four.lvls 1998 1359672.0 0.08710410 1146277.2 1612793.1
#> 3951  four.lvls 1999 1461161.1 0.08870861 1227970.1 1738634.9
#> 3961  four.lvls 2000 1403358.3 0.08703286 1183272.3 1664379.8
#> 3971  four.lvls 2001 1897327.8 0.08585926 1603457.6 2245056.3
#> 3981  four.lvls 2002 2265355.3 0.08509253 1917361.9 2676508.1
#> 3991  four.lvls 2003 2288533.1 0.08114840 1952011.1 2683070.7
#> 4001  four.lvls 2004 2258865.8 0.08140373 1925742.3 2649614.5
#> 4011  four.lvls 2005 2026255.3 0.08630293 1710927.6 2399698.6
#> 4021  four.lvls 2006 1594343.0 0.08722066 1343810.5 1891583.3
#> 4031  four.lvls 2007 1296813.7 0.09097599 1085018.8 1549950.8
#> 4041  four.lvls 2008 1355796.4 0.09267102 1130606.1 1625839.4
#> 4051  four.lvls 2009 1704822.9 0.09432874 1417049.5 2051037.1
#> 4061  four.lvls 2010 1765361.6 0.09991415 1451393.0 2147248.5
#> 4071  four.lvls 2011 2213274.6 0.10336695 1807372.1 2710335.3
#> 4081  four.lvls 2012 2230898.7 0.12615833 1742174.8 2856722.1
#> 387  three.lvls 1992 1074028.7 0.07973987  918628.6 1255717.2
#> 3881 three.lvls 1993  730308.1 0.08616966  616817.9  864679.5
#> 3891 three.lvls 1994  819825.8 0.08662945  691800.9  971543.1
#> 3902 three.lvls 1995  810929.3 0.09269353  676208.5  972490.3
#> 3912 three.lvls 1996  964949.0 0.09515274  800771.7 1162786.6
#> 3922 three.lvls 1997 1191770.0 0.09270028  993766.5 1429224.7
#> 3932 three.lvls 1998 1397638.3 0.08760224 1177135.0 1659446.8
#> 3942 three.lvls 1999 1500111.4 0.08931378 1259209.8 1787100.3
#> 3952 three.lvls 2000 1443703.1 0.08760274 1215931.0 1714142.2
#> 3962 three.lvls 2001 1942545.2 0.08671829 1638909.7 2302434.2
#> 3972 three.lvls 2002 2316718.9 0.08591840 1957663.8 2741628.2
#> 3982 three.lvls 2003 2340002.1 0.08197220 1992691.7 2747846.0
#> 3992 three.lvls 2004 2308481.6 0.08226953 1964704.3 2712412.0
#> 4002 three.lvls 2005 2074764.6 0.08713949 1749017.7 2461180.5
#> 4012 three.lvls 2006 1642330.5 0.08758163 1383278.3 1949896.4
#> 4022 three.lvls 2007 1341867.3 0.09065067 1123430.3 1602776.5
#> 4032 three.lvls 2008 1408215.0 0.09154803 1176905.8 1684985.7
#> 4042 three.lvls 2009 1769008.8 0.09351878 1472737.0 2124881.9
#> 4052 three.lvls 2010 1834220.6 0.09933369 1509722.1 2228466.5
#> 4062 three.lvls 2011 2280443.3 0.10314003 1863050.9 2791347.1
#> 4072 three.lvls 2012 2291249.5 0.12639069 1788489.7 2935339.3
#> 386    two.lvls 1992 1125265.2 0.08505756  952472.4 1329405.1
#> 3871   two.lvls 1993  772373.4 0.09177227  645222.0  924582.0
#> 3882   two.lvls 1994  857414.5 0.09139397  716794.5 1025621.0
#> 3892   two.lvls 1995  847099.7 0.09764166  699552.4 1025767.1
#> 3903   two.lvls 1996 1007184.3 0.10015479  827666.7 1225638.6
#> 3913   two.lvls 1997 1241593.2 0.09743449 1025749.7 1502855.5
#> 3923   two.lvls 1998 1445099.8 0.09279568 1204782.2 1733353.4
#> 3933   two.lvls 1999 1554467.1 0.09457314 1291454.9 1871043.2
#> 3943   two.lvls 2000 1498836.2 0.09327134 1248417.9 1799485.4
#> 3953   two.lvls 2001 1984898.1 0.09148368 1659073.9 2374710.6
#> 3963   two.lvls 2002 2379373.3 0.09090213 1991063.6 2843413.6
#> 3973   two.lvls 2003 2393999.7 0.08706334 2018432.8 2839447.9
#> 3983   two.lvls 2004 2372169.6 0.08693406 2000534.2 2812843.1
#> 3993   two.lvls 2005 2131492.8 0.09196983 1779908.1 2552526.0
#> 4003   two.lvls 2006 1707162.3 0.09235126 1424504.6 2045906.6
#> 4013   two.lvls 2007 1400973.3 0.09505457 1162834.1 1687881.4
#> 4023   two.lvls 2008 1468574.4 0.09632177 1215920.6 1773726.7
#> 4033   two.lvls 2009 1841799.2 0.09851545 1518393.2 2234088.3
#> 4043   two.lvls 2010 1924097.1 0.10502757 1566122.4 2363895.4
#> 4053   two.lvls 2011 2336707.0 0.10786325 1891425.3 2886817.4
#> 4063   two.lvls 2012 2308792.9 0.13110432 1785610.5 2985267.4

rec(HERAS.sams)

#>           name year    value        CV     lbnd      ubnd
#> 187   all.free 1992 54714483 0.1439617 41262887  72551264
#> 196   all.free 1993 45915953 0.1392105 34951450  60320093
#> 205   all.free 1994 34022353 0.1463922 25535997  45328972
#> 214   all.free 1995 50889868 0.1624576 37012177  69970990
#> 223   all.free 1996 40154577 0.1546796 29653031  54375219
#> 232   all.free 1997 23731461 0.1640117 17207386  32729097
#> 241   all.free 1998 16695754 0.1854299 11608202  24013039
#> 250   all.free 1999 62948487 0.1665867 45413383  87254279
#> 259   all.free 2000 44671774 0.1611365 32573999  61262586
#> 268   all.free 2001 81963856 0.1666857 59120308 113633941
#> 277   all.free 2002 40281742 0.1615756 29347581  55289692
#> 286   all.free 2003 19617981 0.1649033 14199916  27103341
#> 295   all.free 2004 27581501 0.1637627 20008766  38020295
#> 304   all.free 2005 23725379 0.1594057 17358984  32426645
#> 313   all.free 2006 27005820 0.1625260 19638687  37136611
#> 322   all.free 2007 26443001 0.1681359 19019127  36764690
#> 331   all.free 2008 31730135 0.1713301 22679464  44392648
#> 340   all.free 2009 42874180 0.1830439 29949232  61377043
#> 349   all.free 2010 35643199 0.2199557 23160428  54853806
#> 358   all.free 2011 46094418 0.3389250 23721903  89566817
#> 367   all.free 2012 45129368 0.6171933 13461508 151295076
#> 185  four.lvls 1992 58256227 0.1323964 44941155  75516262
#> 194  four.lvls 1993 48629556 0.1296185 37719579  62695123
#> 203  four.lvls 1994 34440423 0.1338677 26492198  44773285
#> 212  four.lvls 1995 49908068 0.1443598 37608778  66229626
#> 221  four.lvls 1996 42449639 0.1416641 32157854  56035202
#> 230  four.lvls 1997 27500836 0.1493169 20523187  36850805
#> 239  four.lvls 1998 17280739 0.1579802 12679053  23552542
#> 248  four.lvls 1999 65062170 0.1489855 48585820  87125955
#> 257  four.lvls 2000 45627403 0.1447616 34355971  60596742
#> 266  four.lvls 2001 80473663 0.1483840 60165398 107636793
#> 275  four.lvls 2002 42251812 0.1452270 31785250  56164907
#> 284  four.lvls 2003 19332645 0.1474127 14481417  25809020
#> 293  four.lvls 2004 24840898 0.1486822 18561217  33245137
#> 302  four.lvls 2005 22246933 0.1431073 16805630  29450013
#> 311  four.lvls 2006 26003671 0.1448794 19575410  34542873
#> 320  four.lvls 2007 23618444 0.1579683 17329497  32189676
#> 329  four.lvls 2008 26628066 0.1644025 19292895  36752073
#> 338  four.lvls 2009 36785509 0.1553741 27128116  49880857
#> 347  four.lvls 2010 31415329 0.1688193 22565227  43736449
#> 356  four.lvls 2011 32196642 0.1990350 21796592  47558985
#> 365  four.lvls 2012 27340067 0.3151182 14742310  50702994
#> 184 three.lvls 1992 58760558 0.1337851 45207005  76377615
#> 193 three.lvls 1993 49112100 0.1309828 37992132  63486786
#> 202 three.lvls 1994 34824446 0.1353270 26711084  45402202
#> 211 three.lvls 1995 50650597 0.1453326 38095615  67343262
#> 220 three.lvls 1996 42977696 0.1432006 32459982  56903369
#> 229 three.lvls 1997 27850859 0.1508409 20722409  37431475
#> 238 three.lvls 1998 17432228 0.1594209 12754137  23826198
#> 247 three.lvls 1999 65778221 0.1505103 48973952  88348483
#> 256 three.lvls 2000 46212723 0.1462165 34697611  61549360
#> 265 three.lvls 2001 81298648 0.1498065 60612954 109043854
#> 274 three.lvls 2002 42797380 0.1467043 32102584  57055088
#> 283 three.lvls 2003 19549977 0.1489479 14600215  26177806
#> 292 three.lvls 2004 25099182 0.1501839 18699090  33689817
#> 301 three.lvls 2005 22527473 0.1446703 16965499  29912886
#> 310 three.lvls 2006 26333670 0.1463730 19765881  35083798
#> 319 three.lvls 2007 23819169 0.1594339 17426642  32556634
#> 328 three.lvls 2008 26881022 0.1660744 19412451  37222984
#> 337 three.lvls 2009 37188751 0.1567179 27353357  50560639
#> 346 three.lvls 2010 31758967 0.1698620 22765486  44305313
#> 355 three.lvls 2011 32561348 0.1997808 22011290  48168073
#> 364 three.lvls 2012 27641791 0.3154903 14894138  51299952
#> 183   two.lvls 1992 58939115 0.1366585 45089716  77042384
#> 192   two.lvls 1993 49350355 0.1337860 37967265  64146247
#> 201   two.lvls 1994 34840770 0.1385707 26554247  45713187
#> 210   two.lvls 1995 51236997 0.1476978 38358425  68439459
#> 219   two.lvls 1996 42533132 0.1473627 31863223  56776032
#> 228   two.lvls 1997 27554793 0.1558046 20303627  37395617
#> 237   two.lvls 1998 16875231 0.1647648 12217971  23307752
#> 246   two.lvls 1999 64443356 0.1548987 47569187  87303283
#> 255   two.lvls 2000 45586692 0.1502380 33958844  61196032
#> 264   two.lvls 2001 79377451 0.1539331 58703859 107331612
#> 273   two.lvls 2002 42140028 0.1509491 31347606  56648089
#> 282   two.lvls 2003 19081979 0.1533650 14127864  25773316
#> 291   two.lvls 2004 24620594 0.1548221 18176542  33349228
#> 300   two.lvls 2005 22117740 0.1486007 16529104  29595942
#> 309   two.lvls 2006 25852894 0.1504444 19250780  34719223
#> 318   two.lvls 2007 22991147 0.1641794 16665116  31718522
#> 327   two.lvls 2008 26023803 0.1713021 18601822  36407098
#> 336   two.lvls 2009 35917814 0.1597945 26259718  49128075
#> 345   two.lvls 2010 30450625 0.1719488 21738544  42654215
#> 354   two.lvls 2011 31707284 0.2009254 21385919  47009991
#> 363   two.lvls 2012 27000733 0.3140906 14588688  49972935

fbar(HERAS.sams)

#>            name year      value         CV       lbnd       ubnd
#> 411    all.free 1992 0.41901660 0.09607814 0.34709461 0.50584166
#> 412    all.free 1993 0.51831194 0.09428816 0.43085538 0.62352076
#> 413    all.free 1994 0.48254945 0.09714261 0.39888928 0.58375590
#> 414    all.free 1995 0.48506728 0.10552486 0.39443679 0.59652212
#> 415    all.free 1996 0.23060880 0.11553319 0.18387896 0.28921426
#> 416    all.free 1997 0.20313992 0.11557887 0.16196178 0.25478743
#> 417    all.free 1998 0.25294033 0.11377302 0.20238230 0.31612849
#> 418    all.free 1999 0.22180037 0.11010713 0.17874636 0.27522467
#> 419    all.free 2000 0.23258837 0.10748745 0.18840518 0.28713304
#> 420    all.free 2001 0.19766817 0.10749768 0.16011531 0.24402853
#> 421    all.free 2002 0.18420630 0.10583149 0.14969900 0.22666792
#> 422    all.free 2003 0.20689263 0.10243063 0.16925998 0.25289237
#> 423    all.free 2004 0.24430173 0.10029916 0.20070130 0.29737394
#> 424    all.free 2005 0.31658161 0.09985610 0.26030737 0.38502144
#> 425    all.free 2006 0.26264751 0.10365385 0.21435882 0.32181422
#> 426    all.free 2007 0.23236252 0.11010634 0.18725856 0.28833044
#> 427    all.free 2008 0.14646336 0.11451799 0.11701710 0.18331950
#> 428    all.free 2009 0.07638180 0.11872775 0.06052388 0.09639466
#> 429    all.free 2010 0.07999895 0.11914120 0.06333871 0.10104139
#> 430    all.free 2011 0.09437978 0.12909631 0.07328076 0.12155364
#> 431    all.free 2012 0.09438512 0.37097720 0.04561644 0.19529257
#> 409   four.lvls 1992 0.41701356 0.09485166 0.34626677 0.50221484
#> 410   four.lvls 1993 0.51474005 0.09324890 0.42875865 0.61796379
#> 4111  four.lvls 1994 0.47980093 0.09617957 0.39736662 0.57933637
#> 4121  four.lvls 1995 0.48404770 0.10387085 0.39488580 0.59334162
#> 4131  four.lvls 1996 0.22940187 0.11276614 0.18391133 0.28614452
#> 4141  four.lvls 1997 0.20013320 0.11379913 0.16012213 0.25014217
#> 4151  four.lvls 1998 0.24949110 0.11211454 0.20027246 0.31080564
#> 4161  four.lvls 1999 0.21875861 0.10829866 0.17692103 0.27048976
#> 4171  four.lvls 2000 0.23016471 0.10593754 0.18700916 0.28327914
#> 4181  four.lvls 2001 0.19572100 0.10573838 0.15908568 0.24079294
#> 4191  four.lvls 2002 0.18200375 0.10401640 0.14843618 0.22316233
#> 4201  four.lvls 2003 0.20434767 0.10095343 0.16766267 0.24905942
#> 4211  four.lvls 2004 0.24242222 0.09893066 0.19969213 0.29429569
#> 4221  four.lvls 2005 0.31315845 0.09804884 0.25840641 0.37951155
#> 4231  four.lvls 2006 0.25944927 0.10154877 0.21262405 0.31658658
#> 4241  four.lvls 2007 0.23014238 0.10800521 0.18623475 0.28440190
#> 4251  four.lvls 2008 0.14597523 0.11243959 0.11710317 0.18196576
#> 4261  four.lvls 2009 0.07642197 0.11666994 0.06080045 0.09605715
#> 4271  four.lvls 2010 0.08095568 0.11799289 0.06424063 0.10201990
#> 4281  four.lvls 2011 0.09705495 0.12691977 0.07568004 0.12446692
#> 4291  four.lvls 2012 0.09714144 0.37419360 0.04665353 0.20226677
#> 408  three.lvls 1992 0.40749432 0.09574211 0.33777245 0.49160794
#> 4091 three.lvls 1993 0.50162215 0.09422391 0.41703422 0.60336722
#> 4101 three.lvls 1994 0.46635745 0.09686281 0.38571597 0.56385860
#> 4112 three.lvls 1995 0.46779827 0.10375955 0.38171277 0.57329813
#> 4122 three.lvls 1996 0.22138767 0.11205943 0.17773236 0.27576576
#> 4132 three.lvls 1997 0.19299648 0.11305580 0.15463734 0.24087095
#> 4142 three.lvls 1998 0.24064678 0.11129718 0.19348263 0.29930786
#> 4152 three.lvls 1999 0.21172864 0.10790375 0.17136814 0.26159482
#> 4162 three.lvls 2000 0.22311873 0.10599795 0.18126283 0.27463970
#> 4172 three.lvls 2001 0.19005858 0.10599205 0.15440637 0.23394282
#> 4182 three.lvls 2002 0.17705947 0.10447601 0.14427378 0.21729560
#> 4192 three.lvls 2003 0.19928697 0.10167654 0.16327891 0.24323592
#> 4202 three.lvls 2004 0.23692664 0.09981791 0.19482612 0.28812479
#> 4212 three.lvls 2005 0.30589620 0.09886259 0.25201161 0.37130228
#> 4222 three.lvls 2006 0.25250123 0.10209357 0.20670915 0.30843759
#> 4232 three.lvls 2007 0.22368793 0.10843448 0.18085947 0.27665839
#> 4242 three.lvls 2008 0.14155204 0.11252420 0.11353600 0.17648129
#> 4252 three.lvls 2009 0.07420995 0.11687836 0.05901648 0.09331490
#> 4262 three.lvls 2010 0.07886000 0.11878651 0.06248038 0.09953364
#> 4272 three.lvls 2011 0.09500397 0.12806567 0.07391457 0.12211062
#> 4282 three.lvls 2012 0.09508764 0.37579887 0.04552371 0.19861432
#> 407    two.lvls 1992 0.39202415 0.09908492 0.32282714 0.47605334
#> 4081   two.lvls 1993 0.47993896 0.09827857 0.39584905 0.58189204
#> 4092   two.lvls 1994 0.44486241 0.10067949 0.36519565 0.54190832
#> 4102   two.lvls 1995 0.44312297 0.10731569 0.35906684 0.54685630
#> 4113   two.lvls 1996 0.20939588 0.11483063 0.16719464 0.26224906
#> 4123   two.lvls 1997 0.18249539 0.11583147 0.14543005 0.22900747
#> 4133   two.lvls 1998 0.22789642 0.11383405 0.18232239 0.28486232
#> 4143   two.lvls 1999 0.20188497 0.11045741 0.16258510 0.25068435
#> 4153   two.lvls 2000 0.21310783 0.10899491 0.17211593 0.26386254
#> 4163   two.lvls 2001 0.18190507 0.10891469 0.14693821 0.22519299
#> 4173   two.lvls 2002 0.16972377 0.10738155 0.13751107 0.20948247
#> 4183   two.lvls 2003 0.19165750 0.10447522 0.15616896 0.23521062
#> 4193   two.lvls 2004 0.22847844 0.10238173 0.18693737 0.27925073
#> 4203   two.lvls 2005 0.29550023 0.10130082 0.24228628 0.36040169
#> 4213   two.lvls 2006 0.24313658 0.10465637 0.19804551 0.29849400
#> 4223   two.lvls 2007 0.21546654 0.11080704 0.17340395 0.26773226
#> 4233   two.lvls 2008 0.13564275 0.11515152 0.10823747 0.16998694
#> 4243   two.lvls 2009 0.07096232 0.11965131 0.05612787 0.08971748
#> 4253   two.lvls 2010 0.07574590 0.12216273 0.05961727 0.09623790
#> 4263   two.lvls 2011 0.09205044 0.13127204 0.07116802 0.11906028
#> 4273   two.lvls 2012 0.09212981 0.37861244 0.04386507 0.19350028