Accessor functions for FLSAM

Returns results from the FLSAM asessment model

catchabilities(object)
power.law.exps(object)

obs.var(object)

ssb(object)
tsb(object)

n(object)
rec(object)

f(object)
fbar(object)

catch(object)

Arguments

object

An FLSAM or FLSAMs :object.

Details

Extracting many of the fitted values directly from the parameter slot of an FLSAM object can lead to difficulties in interpretation, due to the binding of parameters (i.e. fitting one parameter to multiple time series) and the fitting of parameters in log-space. These accessor functions provide easy access to these key results, correcting for these issues and providing both estimates of the parameter concerned and the associated confidence intervals.

The following accessor functions are currently supported:

catchabilities - linear proportionality parameter relating the observations (e.g. survey indices) to the modelled values (e.g. numbers at age, SSB)

power.law.exps in cases where the relationship between the observations and the modelled values is modelled as being a power-law, this function returns the associated coefficients

obs.var - the variances estimated internally within the model that are associated with each individual time series. These variances are mathematically analogous to weighting factors

ssb - the spawning stock biomass estimated internally within the model

tsb - the total stock biomass estimated internally within the model

n - the estimate number of individuals at a given age in a given year

rec - the estimated recruitment of fish at the youngest modelled age

f - the fishing mortality on a given age in a given year

fbar - the mean fishing mortality in a given year

catch - the (modelled) total catch of fish (in weight) in a given year

Accessor functions are provided for both FLSAM and FLSAMs objects.

Value

The accessor functions return a data.frame whoose structure is dependent on the function called. Generally the data.frame can be considered to have two parts.

The first part is the meta-data part, which gives context-specific information about the contents of each row. Possible columns returned are

name - In cases where an FLSAMs object is supplied, the name of the corresponding object is stored in this column. There is no requirement that FLSAMs object have unique names - however, extracting and processing data can be challenging if they do not!

fleet - The name of the fleet that the parameter corresponds to

year - The year that the parameter corresponds to

age - The age that the parameter corresponds to. In the case of biomass (or other non-age-based observations), age is NA.

The second part of the data-frame is common and constant to all accessor functions, and contains the actual parameters and their uncertainties

value - The value of the parameter

CV - The coefficient of variation for the parameter estimate i.e. standard deviation of the log-transformed parameter

ubnd - The upper bound (95% confidence limit) for the parameter estimate

lbnd - The lower bound (95% confidence limit) for the parameter estimate

Note

The ssb, tsb, fbar and catch accessor functions read the values estimated directly within the model. However, the estimation of these parameters is dependent on the calculation of the hessian: when the FLSAM control object has the nohess option set to true, however, the hessian will not be estimated. In these cases, these function calls will fail with a warning. If you still want to calculate the values in this case, update the corresponding stock object: see the examples.

See also

FLSAM, params, coef, coefficients

Examples

#Load data
library(FLSAM)
data(NSH.sam)
#Extract parameter
catchabilities(NSH.sam)

#>      fleet age        value       CV         lbnd         ubnd
#> 20    SCAI  NA 1.277455e-02 0.137280 9.760918e-03 1.671864e-02
#> 3    HERAS   1 9.743546e-01 0.073280 8.439949e-01 1.124849e+00
#> 4    HERAS   2 9.743546e-01 0.073280 8.439949e-01 1.124849e+00
#> 5    HERAS   3 1.152922e+00 0.069755 1.005596e+00 1.321834e+00
#> 6    HERAS   4 1.152922e+00 0.069755 1.005596e+00 1.321834e+00
#> 7    HERAS   5 1.307543e+00 0.094497 1.086472e+00 1.573597e+00
#> 8    HERAS   6 1.307543e+00 0.094497 1.086472e+00 1.573597e+00
#> 9    HERAS   7 1.307543e+00 0.094497 1.086472e+00 1.573597e+00
#> 10   HERAS   8 1.307543e+00 0.094497 1.086472e+00 1.573597e+00
#> 1  IBTS-Q1   1 1.555256e-04 0.089933 1.303916e-04 1.855045e-04
#> 2    IBTS0   0 3.081790e-06 0.102280 2.521974e-06 3.765872e-06

obs.var(NSH.sam)

#>      fleet age      value         CV        lbnd       ubnd
#> 3    catch   0 0.04978707 8.0482e-06 0.049786283 0.04978785
#> 4    catch   1 0.09130954 1.1573e+00 0.009449363 0.88232748
#> 6    catch   2 0.09130954 1.1573e+00 0.009449363 0.88232748
#> 5    catch   3 0.09130954 1.1573e+00 0.009449363 0.88232748
#> 8    catch   4 0.09130954 1.1573e+00 0.009449363 0.88232748
#> 7    catch   5 0.09130954 1.1573e+00 0.009449363 0.88232748
#> 9    catch   6 0.23976404 1.9074e-01 0.164976803 0.34845381
#> 10   catch   7 0.23976404 1.9074e-01 0.164976803 0.34845381
#> 11   catch   8 0.23976404 1.9074e-01 0.164976803 0.34845381
#> 21    SCAI  NA 0.54014373 1.6250e-01 0.392813631 0.74273200
#> 12   HERAS   1 0.39266439 2.0510e-01 0.262685968 0.58695683
#> 13   HERAS   2 0.20990505 1.1509e-01 0.167516002 0.26302042
#> 14   HERAS   3 0.20990505 1.1509e-01 0.167516002 0.26302042
#> 15   HERAS   4 0.20990505 1.1509e-01 0.167516002 0.26302042
#> 16   HERAS   5 0.20990505 1.1509e-01 0.167516002 0.26302042
#> 17   HERAS   6 0.27296819 1.5107e-01 0.203010577 0.36703326
#> 18   HERAS   7 0.27296819 1.5107e-01 0.203010577 0.36703326
#> 19   HERAS   8 0.27296819 1.5107e-01 0.203010577 0.36703326
#> 1  IBTS-Q1   1 0.30620514 1.8537e-01 0.212922889 0.44035467
#> 2    IBTS0   0 0.35743563 1.8818e-01 0.247181400 0.51686829

ssb(NSH.sam)

#>     year     value       CV      lbnd      ubnd
#> 400 1992 1073033.2 0.077939  921022.3 1250132.9
#> 401 1993  747881.6 0.085889  632008.1  884999.6
#> 402 1994  817494.9 0.082786  695050.3  961510.3
#> 403 1995  852561.0 0.089104  715943.2 1015248.4
#> 404 1996  960299.0 0.089875  805199.2 1145274.6
#> 405 1997 1102399.8 0.090157  923838.2 1315474.1
#> 406 1998 1358647.1 0.083072 1154501.1 1598891.5
#> 407 1999 1441219.2 0.083246 1224248.6 1696643.0
#> 408 2000 1431165.9 0.082359 1217824.2 1681881.5
#> 409 2001 1905014.2 0.082330 1621128.7 2238612.5
#> 410 2002 2231088.2 0.080739 1904540.9 2613624.3
#> 411 2003 2280716.0 0.077626 1958820.5 2655509.1
#> 412 2004 2264806.8 0.078647 1941267.9 2642267.8
#> 413 2005 2156506.3 0.081752 1837223.6 2531275.6
#> 414 2006 1677810.1 0.083852 1423529.7 1977511.7
#> 415 2007 1349169.8 0.086860 1137967.6 1599570.3
#> 416 2008 1381941.6 0.087412 1164348.8 1640198.0
#> 417 2009 1667773.4 0.090396 1396980.7 1991056.9
#> 418 2010 1741051.8 0.100490 1429792.1 2120071.6
#> 419 2011 2101159.7 0.098890 1730940.7 2550562.3
#> 420 2012 2186909.7 0.122880 1718831.4 2782456.7

fbar(NSH.sam)

#>     year      value       CV       lbnd      ubnd
#> 463 1992 0.39991230 0.091378 0.33433531 0.4783517
#> 464 1993 0.45589170 0.083983 0.38669977 0.5374641
#> 465 1994 0.47078206 0.083762 0.39950319 0.5547784
#> 466 1995 0.42212660 0.107930 0.34164168 0.5215724
#> 467 1996 0.23747332 0.130530 0.18386770 0.3067074
#> 468 1997 0.21612464 0.109540 0.17436606 0.2678839
#> 469 1998 0.23873527 0.097885 0.19705851 0.2892264
#> 470 1999 0.22535012 0.094808 0.18713525 0.2713688
#> 471 2000 0.23084703 0.094680 0.19174809 0.2779185
#> 472 2001 0.19970780 0.095563 0.16559612 0.2408462
#> 473 2002 0.18626219 0.098146 0.15366718 0.2257711
#> 474 2003 0.21521882 0.092134 0.17966116 0.2578139
#> 475 2004 0.25357389 0.094634 0.21064466 0.3052521
#> 476 2005 0.27395265 0.097094 0.22647873 0.3313779
#> 477 2006 0.24468099 0.095830 0.20278139 0.2952381
#> 478 2007 0.21816579 0.111730 0.17525893 0.2715771
#> 479 2008 0.14291575 0.102490 0.11690655 0.1747114
#> 480 2009 0.08191280 0.139450 0.06232317 0.1076599
#> 481 2010 0.08671746 0.110270 0.06986229 0.1076392
#> 482 2011 0.10440267 0.111240 0.08395029 0.1298378
#> 483 2012 0.10440267 0.184920 0.07266153 0.1500095

rec(NSH.sam)

#>     year    value      CV     lbnd      ubnd
#> 22  1992 57483011 0.11856 45563723  72520338
#> 39  1993 49278577 0.11542 39301640  61788214
#> 56  1994 34072460 0.12962 26428269  43927680
#> 73  1995 44767814 0.12842 34805877  57580999
#> 90  1996 41533048 0.14443 31293382  55123286
#> 107 1997 29092763 0.15267 21568982  39241021
#> 124 1998 22454464 0.16225 16337767  30861191
#> 141 1999 64038819 0.15593 47175120  86930786
#> 158 2000 46641441 0.15069 34713791  62667429
#> 175 2001 77593949 0.15943 56769922 106056529
#> 192 2002 41037630 0.15043 30558615  55110058
#> 209 2003 20957350 0.15572 15444891  28437271
#> 226 2004 24227526 0.15127 18011316  32589125
#> 243 2005 21965860 0.14700 16467182  29300642
#> 260 2006 25418810 0.14848 19000564  34005091
#> 277 2007 24692250 0.16505 17867643  34123539
#> 294 2008 25393404 0.16932 18221865  35387429
#> 311 2009 32508105 0.16167 23679655  44628053
#> 328 2010 33498124 0.18751 23195768  48376252
#> 345 2011 29326438 0.20556 19601221  43876859
#> 362 2012 26722060 0.29907 14869502  48022355

#Power law parameters could be extracted as follows, but this will
#return an error in this case, as there are no power-law exponents in the model
#power.law.exps(NSH.sam) 

#And for FLSAMs
data(HERAS.sams)
ssb(HERAS.sams)

#>            name year     value       CV      lbnd      ubnd
#> 405    all.free 1992  990534.5 0.076689  852296.3 1151194.3
#> 406    all.free 1993  674684.2 0.087501  568353.0  800908.5
#> 407    all.free 1994  745641.4 0.084258  632132.5  879532.4
#> 408    all.free 1995  779961.6 0.091036  652502.0  932319.2
#> 409    all.free 1996  885581.7 0.090577  741528.7 1057619.2
#> 410    all.free 1997 1017643.3 0.091491  850583.3 1217514.9
#> 411    all.free 1998 1269330.3 0.082185 1080481.6 1491186.4
#> 412    all.free 1999 1353223.4 0.081573 1153275.7 1587836.7
#> 413    all.free 2000 1334410.3 0.081430 1137561.2 1565323.1
#> 414    all.free 2001 1803067.6 0.080216 1540745.0 2110052.4
#> 415    all.free 2002 2118036.4 0.078738 1815140.7 2471476.9
#> 416    all.free 2003 2162985.5 0.076228 1862803.4 2511540.5
#> 417    all.free 2004 2145750.6 0.077583 1843059.1 2498154.2
#> 418    all.free 2005 2030921.1 0.080889 1733161.2 2379836.7
#> 419    all.free 2006 1556577.4 0.084307 1319493.2 1836260.5
#> 420    all.free 2007 1241710.0 0.087968 1045057.8 1475366.9
#> 421    all.free 2008 1262999.5 0.089248 1060312.2 1504432.2
#> 422    all.free 2009 1530339.2 0.091154 1279958.4 1829698.7
#> 423    all.free 2010 1580102.1 0.101830 1294212.8 1929143.7
#> 424    all.free 2011 1951287.6 0.099469 1605652.5 2371324.6
#> 425    all.free 2012 2076096.5 0.122560 1632760.0 2639810.3
#> 401   four.lvls 1992 1034056.5 0.079023  885683.5 1207285.6
#> 402   four.lvls 1993  713544.4 0.089125  599178.6  849739.4
#> 403   four.lvls 1994  783871.2 0.085387  663073.7  926675.3
#> 404   four.lvls 1995  819131.6 0.091402  684779.4  979843.3
#> 4051  four.lvls 1996  926343.1 0.091191  774726.7 1107631.3
#> 4061  four.lvls 1997 1064483.2 0.091526  889672.8 1273641.9
#> 4071  four.lvls 1998 1317175.2 0.083568 1118173.1 1551593.9
#> 4081  four.lvls 1999 1400024.1 0.083266 1189208.6 1648211.6
#> 4091  four.lvls 2000 1386093.6 0.082759 1178546.4 1630190.8
#> 4101  four.lvls 2001 1856122.1 0.081873 1580938.0 2179205.9
#> 4111  four.lvls 2002 2176002.4 0.080272 1859218.7 2546761.5
#> 4121  four.lvls 2003 2224405.0 0.077342 1911520.7 2588503.1
#> 4131  four.lvls 2004 2208888.5 0.078390 1894291.8 2575732.2
#> 4141  four.lvls 2005 2099059.6 0.081540 1789025.5 2462822.0
#> 4151  four.lvls 2006 1621723.5 0.084517 1374151.0 1913899.5
#> 4161  four.lvls 2007 1298863.2 0.088431 1092168.0 1544676.0
#> 4171  four.lvls 2008 1323777.6 0.090403 1108823.5 1580402.1
#> 4181  four.lvls 2009 1599177.5 0.092800 1333225.8 1918181.2
#> 4191  four.lvls 2010 1662777.6 0.103280 1358064.7 2035859.8
#> 4201  four.lvls 2011 2022813.7 0.100700 1660497.8 2464185.8
#> 4211  four.lvls 2012 2115919.4 0.123440 1661211.3 2695090.7
#> 400  three.lvls 1992 1073033.2 0.077939  921022.3 1250132.9
#> 4011 three.lvls 1993  747881.6 0.085889  632008.1  884999.6
#> 4021 three.lvls 1994  817494.9 0.082786  695050.3  961510.3
#> 4031 three.lvls 1995  852561.0 0.089104  715943.2 1015248.4
#> 4041 three.lvls 1996  960299.0 0.089875  805199.2 1145274.6
#> 4052 three.lvls 1997 1102399.8 0.090157  923838.2 1315474.1
#> 4062 three.lvls 1998 1358647.1 0.083072 1154501.1 1598891.5
#> 4072 three.lvls 1999 1441219.2 0.083246 1224248.6 1696643.0
#> 4082 three.lvls 2000 1431165.9 0.082359 1217824.2 1681881.5
#> 4092 three.lvls 2001 1905014.2 0.082330 1621128.7 2238612.5
#> 4102 three.lvls 2002 2231088.2 0.080739 1904540.9 2613624.3
#> 4112 three.lvls 2003 2280716.0 0.077626 1958820.5 2655509.1
#> 4122 three.lvls 2004 2264806.8 0.078647 1941267.9 2642267.8
#> 4132 three.lvls 2005 2156506.3 0.081752 1837223.6 2531275.6
#> 4142 three.lvls 2006 1677810.1 0.083852 1423529.7 1977511.7
#> 4152 three.lvls 2007 1349169.8 0.086860 1137967.6 1599570.3
#> 4162 three.lvls 2008 1381941.6 0.087412 1164348.8 1640198.0
#> 4172 three.lvls 2009 1667773.4 0.090396 1396980.7 1991056.9
#> 4182 three.lvls 2010 1741051.8 0.100490 1429792.1 2120071.6
#> 4192 three.lvls 2011 2101159.7 0.098890 1730940.7 2550562.3
#> 4202 three.lvls 2012 2186909.7 0.122880 1718831.4 2782456.7
#> 399    two.lvls 1992 1130307.2 0.082555  961444.6 1328827.9
#> 4001   two.lvls 1993  796514.0 0.090037  667655.4  950242.5
#> 4012   two.lvls 1994  858549.8 0.086589  724535.2 1017352.7
#> 4022   two.lvls 1995  893587.9 0.092710  745110.9 1071651.6
#> 4032   two.lvls 1996 1000489.6 0.094172  831862.9 1203298.5
#> 4042   two.lvls 1997 1148537.6 0.094265  954784.2 1381609.1
#> 4053   two.lvls 1998 1409858.6 0.087746 1187092.8 1674428.0
#> 4063   two.lvls 1999 1498537.2 0.088037 1261040.2 1780763.1
#> 4073   two.lvls 2000 1491063.2 0.087400 1256318.3 1769670.6
#> 4083   two.lvls 2001 1955194.0 0.086730 1649543.6 2317479.7
#> 4093   two.lvls 2002 2296737.1 0.085471 1942481.5 2715599.1
#> 4103   two.lvls 2003 2340792.3 0.082374 1991795.3 2750939.5
#> 4113   two.lvls 2004 2338452.6 0.082750 1988338.7 2750216.0
#> 4123   two.lvls 2005 2233320.4 0.086288 1885824.2 2644848.9
#> 4133   two.lvls 2006 1753282.0 0.087748 1476247.5 2082305.1
#> 4143   two.lvls 2007 1415509.4 0.090303 1185892.4 1689585.7
#> 4153   two.lvls 2008 1451343.2 0.090817 1214689.1 1734103.8
#> 4163   two.lvls 2009 1746282.8 0.094123 1452096.5 2100069.7
#> 4173   two.lvls 2010 1846864.7 0.104450 1504961.7 2266442.5
#> 4183   two.lvls 2011 2180358.8 0.102590 1783206.3 2665964.3
#> 4193   two.lvls 2012 2233320.4 0.126860 1741669.0 2863758.8

rec(HERAS.sams)

#>           name year    value      CV     lbnd      ubnd
#> 27    all.free 1992 56232194 0.11168 45177386  69992090
#> 44    all.free 1993 48013847 0.10916 38765700  59468279
#> 61    all.free 1994 33032418 0.12361 25925158  42088100
#> 78    all.free 1995 43314586 0.12314 34026344  55138260
#> 95    all.free 1996 41078688 0.13826 31327611  53864899
#> 112   all.free 1997 28261189 0.14709 21182874  37704741
#> 129   all.free 1998 22098052 0.15485 16313344  29934014
#> 146   all.free 1999 63592112 0.14936 47453194  85219906
#> 163   all.free 2000 45489860 0.14553 34200855  60505135
#> 180   all.free 2001 77904946 0.15433 57570060 105422516
#> 197   all.free 2002 40225030 0.14495 30276973  53441705
#> 214   all.free 2003 20665992 0.14923 15425140  27687478
#> 231   all.free 2004 24034479 0.14485 18094055  31925193
#> 248   all.free 2005 21509387 0.14223 16276440  28424750
#> 265   all.free 2006 25015345 0.14308 18897935  33113010
#> 282   all.free 2007 24716955 0.15810 18130823  33695539
#> 299   all.free 2008 25393404 0.16197 18486269  34881292
#> 316   all.free 2009 32475613 0.15844 23806224  44302089
#> 333   all.free 2010 33970396 0.18648 23570328  48959344
#> 350   all.free 2011 29092763 0.20466 19479368  43450530
#> 367   all.free 2012 26456171 0.30165 14647292  47785557
#> 23   four.lvls 1992 56911045 0.11560 45372829  71383407
#> 40   four.lvls 1993 48690768 0.11270 39040417  60726577
#> 57   four.lvls 1994 33598769 0.12696 26197077  43091728
#> 74   four.lvls 1995 44057229 0.12598 34417620  56396677
#> 91   four.lvls 1996 41078688 0.14190 31104902  54250567
#> 108  four.lvls 1997 28602366 0.15025 21306228  38397004
#> 125  four.lvls 1998 22164445 0.15929 16220584  30286372
#> 142  four.lvls 1999 63338252 0.15309 46919485  85502519
#> 159  four.lvls 2000 45855238 0.14844 34279484  61339981
#> 176  four.lvls 2001 77129779 0.15729 56667511 104980838
#> 193  four.lvls 2002 40386252 0.14807 30212998  53985022
#> 210  four.lvls 2003 20665992 0.15310 15308580  27898291
#> 227  four.lvls 2004 24010457 0.14878 17937269  32139900
#> 244  four.lvls 2005 21595597 0.14472 16262116  28678297
#> 261  four.lvls 2006 25065426 0.14609 18824384  33375625
#> 278  four.lvls 2007 24520008 0.16242 17834704  33711285
#> 295  four.lvls 2008 25140735 0.16632 18146945  34829915
#> 312  four.lvls 2009 32088235 0.15980 23459639  43890480
#> 329  four.lvls 2010 33131664 0.18657 22984320  47758958
#> 346  four.lvls 2011 28976624 0.20494 19390961  43300832
#> 363  four.lvls 2012 26403312 0.29933 14684649  47473717
#> 22  three.lvls 1992 57483011 0.11856 45563723  72520338
#> 39  three.lvls 1993 49278577 0.11542 39301640  61788214
#> 56  three.lvls 1994 34072460 0.12962 26428269  43927680
#> 73  three.lvls 1995 44767814 0.12842 34805877  57580999
#> 90  three.lvls 1996 41533048 0.14443 31293382  55123286
#> 107 three.lvls 1997 29092763 0.15267 21568982  39241021
#> 124 three.lvls 1998 22454464 0.16225 16337767  30861191
#> 141 three.lvls 1999 64038819 0.15593 47175120  86930786
#> 158 three.lvls 2000 46641441 0.15069 34713791  62667429
#> 175 three.lvls 2001 77593949 0.15943 56769922 106056529
#> 192 three.lvls 2002 41037630 0.15043 30558615  55110058
#> 209 three.lvls 2003 20957350 0.15572 15444891  28437271
#> 226 three.lvls 2004 24227526 0.15127 18011316  32589125
#> 243 three.lvls 2005 21965860 0.14700 16467182  29300642
#> 260 three.lvls 2006 25418810 0.14848 19000564  34005091
#> 277 three.lvls 2007 24692250 0.16505 17867643  34123539
#> 294 three.lvls 2008 25393404 0.16932 18221865  35387429
#> 311 three.lvls 2009 32508105 0.16167 23679655  44628053
#> 328 three.lvls 2010 33498124 0.18751 23195768  48376252
#> 345 three.lvls 2011 29326438 0.20556 19601221  43876859
#> 362 three.lvls 2012 26722060 0.29907 14869502  48022355
#> 21    two.lvls 1992 57886804 0.12464 45340245  73905248
#> 38    two.lvls 1993 49674387 0.12065 39213279  62926252
#> 55    two.lvls 1994 34346133 0.13481 26370919  44733249
#> 72    two.lvls 1995 45127393 0.13304 34769169  58571477
#> 89    two.lvls 1996 41202109 0.15034 30686507  55321180
#> 106   two.lvls 1997 29326438 0.15853 21493932  40013154
#> 123   two.lvls 1998 22364826 0.16973 16035719  31191956
#> 140   two.lvls 1999 63148522 0.16217 45953786  86777090
#> 157   two.lvls 2000 46734817 0.15604 34420456  63454800
#> 174   two.lvls 2001 75753863 0.16498 54824032 104673946
#> 191   two.lvls 2002 40996613 0.15614 30188318  55674590
#> 208   two.lvls 2003 20852825 0.16235 15169449  28665532
#> 225   two.lvls 2004 23914607 0.15778 17553277  32581291
#> 242   two.lvls 2005 21921972 0.15197 16274968  29528345
#> 259   two.lvls 2006 25266754 0.15409 18680366  34175392
#> 276   two.lvls 2007 24082596 0.17223 17182968  33752694
#> 293   two.lvls 2008 24816021 0.17726 17532562  35125208
#> 310   two.lvls 2009 31578908 0.16525 22841966  43657690
#> 327   two.lvls 2010 32605776 0.19032 22453853  47347624
#> 344   two.lvls 2011 29005615 0.20661 19346932  43486261
#> 361   two.lvls 2012 26482640 0.29686 14800247  47386389

fbar(HERAS.sams)

#>            name year      value       CV       lbnd      ubnd
#> 468    all.free 1992 0.42619853 0.090217 0.35712289 0.5086350
#> 469    all.free 1993 0.48789259 0.082321 0.41519410 0.5733202
#> 470    all.free 1994 0.50625742 0.084001 0.42940620 0.5968628
#> 471    all.free 1995 0.45997644 0.108540 0.37183004 0.5690189
#> 472    all.free 1996 0.26350050 0.136950 0.20146855 0.3446320
#> 473    all.free 1997 0.23849665 0.116850 0.18967826 0.2998797
#> 474    all.free 1998 0.26221251 0.104520 0.21364080 0.3218271
#> 475    all.free 1999 0.24568624 0.101860 0.20122218 0.2999755
#> 476    all.free 2000 0.24974872 0.097961 0.20611860 0.3026142
#> 477    all.free 2001 0.21603821 0.098917 0.17796341 0.2622590
#> 478    all.free 2002 0.20014764 0.100900 0.16423384 0.2439149
#> 479    all.free 2003 0.22939726 0.094276 0.19069481 0.2759546
#> 480    all.free 2004 0.26876981 0.094803 0.22319401 0.3236521
#> 481    all.free 2005 0.28964478 0.096269 0.23983904 0.3497933
#> 482    all.free 2006 0.26048761 0.095625 0.21596801 0.3141845
#> 483    all.free 2007 0.23191137 0.110250 0.18684236 0.2878517
#> 484    all.free 2008 0.15327831 0.102860 0.12529233 0.1875154
#> 485    all.free 2009 0.08879721 0.140580 0.06741169 0.1169670
#> 486    all.free 2010 0.09311686 0.111650 0.07481522 0.1158955
#> 487    all.free 2011 0.11019541 0.110750 0.08869338 0.1369102
#> 488    all.free 2012 0.11019541 0.183360 0.07692798 0.1578493
#> 464   four.lvls 1992 0.41197608 0.091954 0.34403226 0.4933383
#> 465   four.lvls 1993 0.47042441 0.084737 0.39843754 0.5554173
#> 466   four.lvls 1994 0.48635328 0.085213 0.41154476 0.5747601
#> 467   four.lvls 1995 0.43827444 0.109370 0.35371096 0.5430549
#> 4681  four.lvls 1996 0.24837887 0.134010 0.19100425 0.3229879
#> 4691  four.lvls 1997 0.22559814 0.113740 0.18051699 0.2819376
#> 4701  four.lvls 1998 0.24865224 0.101660 0.20373124 0.3034779
#> 4711  four.lvls 1999 0.23384425 0.098076 0.19294908 0.2834071
#> 4721  four.lvls 2000 0.23887855 0.096541 0.19769687 0.2886387
#> 4731  four.lvls 2001 0.20675929 0.097439 0.17081393 0.2502688
#> 4741  four.lvls 2002 0.19228051 0.099508 0.15820942 0.2336890
#> 4751  four.lvls 2003 0.22128584 0.092993 0.18441506 0.2655283
#> 4761  four.lvls 2004 0.26025327 0.094704 0.21616358 0.3133357
#> 4771  four.lvls 2005 0.28102827 0.096921 0.23240699 0.3398215
#> 4781  four.lvls 2006 0.25185544 0.096217 0.20856901 0.3041255
#> 4791  four.lvls 2007 0.22467508 0.112040 0.18037840 0.2798500
#> 4801  four.lvls 2008 0.14781408 0.104120 0.12052776 0.1812778
#> 4811  four.lvls 2009 0.08505136 0.140850 0.06453381 0.1120921
#> 4821  four.lvls 2010 0.08958208 0.111600 0.07198223 0.1114851
#> 4831  four.lvls 2011 0.10701353 0.111110 0.08607161 0.1330508
#> 4841  four.lvls 2012 0.10701353 0.184060 0.07460427 0.1535019
#> 463  three.lvls 1992 0.39991230 0.091378 0.33433531 0.4783517
#> 4641 three.lvls 1993 0.45589170 0.083983 0.38669977 0.5374641
#> 4651 three.lvls 1994 0.47078206 0.083762 0.39950319 0.5547784
#> 4661 three.lvls 1995 0.42212660 0.107930 0.34164168 0.5215724
#> 4671 three.lvls 1996 0.23747332 0.130530 0.18386770 0.3067074
#> 4682 three.lvls 1997 0.21612464 0.109540 0.17436606 0.2678839
#> 4692 three.lvls 1998 0.23873527 0.097885 0.19705851 0.2892264
#> 4702 three.lvls 1999 0.22535012 0.094808 0.18713525 0.2713688
#> 4712 three.lvls 2000 0.23084703 0.094680 0.19174809 0.2779185
#> 4722 three.lvls 2001 0.19970780 0.095563 0.16559612 0.2408462
#> 4732 three.lvls 2002 0.18626219 0.098146 0.15366718 0.2257711
#> 4742 three.lvls 2003 0.21521882 0.092134 0.17966116 0.2578139
#> 4752 three.lvls 2004 0.25357389 0.094634 0.21064466 0.3052521
#> 4762 three.lvls 2005 0.27395265 0.097094 0.22647873 0.3313779
#> 4772 three.lvls 2006 0.24468099 0.095830 0.20278139 0.2952381
#> 4782 three.lvls 2007 0.21816579 0.111730 0.17525893 0.2715771
#> 4792 three.lvls 2008 0.14291575 0.102490 0.11690655 0.1747114
#> 4802 three.lvls 2009 0.08191280 0.139450 0.06232317 0.1076599
#> 4812 three.lvls 2010 0.08671746 0.110270 0.06986229 0.1076392
#> 4822 three.lvls 2011 0.10440267 0.111240 0.08395029 0.1298378
#> 4832 three.lvls 2012 0.10440267 0.184920 0.07266153 0.1500095
#> 462    two.lvls 1992 0.38261348 0.094795 0.31773798 0.4607352
#> 4631   two.lvls 1993 0.43536958 0.087747 0.36657794 0.5170706
#> 4642   two.lvls 1994 0.44917622 0.087561 0.37834093 0.5332737
#> 4652   two.lvls 1995 0.39934084 0.111430 0.32099081 0.4968152
#> 4662   two.lvls 1996 0.22402447 0.132290 0.17285738 0.2903374
#> 4672   two.lvls 1997 0.20419089 0.111360 0.16415148 0.2539966
#> 4683   two.lvls 1998 0.22627595 0.099880 0.18604535 0.2752060
#> 4693   two.lvls 1999 0.21455267 0.096473 0.17758834 0.2592110
#> 4703   two.lvls 2000 0.22002810 0.097293 0.18182797 0.2662537
#> 4713   two.lvls 2001 0.19027212 0.098512 0.15686283 0.2307971
#> 4723   two.lvls 2002 0.17822651 0.100800 0.14627483 0.2171576
#> 4733   two.lvls 2003 0.20634619 0.094293 0.17152705 0.2482334
#> 4743   two.lvls 2004 0.24463206 0.097200 0.20219716 0.2959727
#> 4753   two.lvls 2005 0.26471540 0.099436 0.21783997 0.3216776
#> 4763   two.lvls 2006 0.23555756 0.098389 0.19424356 0.2856587
#> 4773   two.lvls 2007 0.20936000 0.113500 0.16760253 0.2615212
#> 4783   two.lvls 2008 0.13613612 0.103910 0.11105124 0.1668873
#> 4793   two.lvls 2009 0.07774664 0.142290 0.05882500 0.1027546
#> 4803   two.lvls 2010 0.08254596 0.112900 0.06615968 0.1029908
#> 4813   two.lvls 2011 0.10057013 0.113920 0.08044487 0.1257302
#> 4823   two.lvls 2012 0.10057013 0.186690 0.06975178 0.1450049

#Handling "nohess" situation
data(NSH)
ctrl.nohess <- NSH.ctrl
ctrl.nohess@nohess <- TRUE
sam.nohess <- FLSAM(NSH,NSH.tun,ctrl.nohess)

#> Error in attributes(x) <- `*vtmp*`: length of 'dimnames' [1] not equal to array extent

stck.nohess <- NSH + sam.nohess

#> Error in eval(expr, envir, enclos): object 'sam.nohess' not found

plot(stck.nohess)

#> Error in plot(stck.nohess): object 'stck.nohess' not found

ssb(stck.nohess)

#> Error in ssb(stck.nohess): object 'stck.nohess' not found