Perform a sequence of assessments where each of the fleets are excluded or included from the assessment run. It is also possible to run intermediate states where 2 or more fleets are excluded or included from the assessment

looi(stck, tun, ctrl, type="full")

loo(stck, tun, ctrl)
loi(stck, tun, ctrl)

Arguments

stck

FLSAM object that specifies the stock characteristics

tun

FLIndices object to use within the 'looi' assessment and out of which fleets will be included or dropped

ctrl

FLSAM.control object that specifies the assessment settings

type

Selects the type of LOOI run to perform. Options are

type: 'full' runs all combinations of leaving fleets in and out (may take a long time to run)

type: 'LOI' runs all combinations keeping only 1 fleet in

type: 'LOO' runs all combinations leaving only 1 fleet out.

Default value is 'full'.

Details

The 'loo' and 'loi' methods call the 'looi' method with the appropriate type selected, and can be used as convenient shotcut

Value

An FLSAMs object is returned where the names slot contains the combination of fleets used within the assessments

See also

FLSAMs

Examples

#- Load data
data(NSH)
data(NSH.sam)

#- Run a LOI
one.in <- loi(NSH,NSH.tun,NSH.ctrl)

#> Running LOI analyses

#> Error in attributes(x) <- `*vtmp*`: length of 'dimnames' [1] not equal to array extent

plot(one.in)

#> Error in plot(one.in): object 'one.in' not found

#- Run a LOO
one.out <- loo(NSH,NSH.tun,NSH.ctrl)

#> Running LOO analyses

#> Error in attributes(x) <- `*vtmp*`: length of 'dimnames' [1] not equal to array extent

plot(one.out)

#> Error in plot(one.out): object 'one.out' not found